[add]viewModelを追加しUI 状態の管理はviewModelで行うよう変更
