+++ app/src/main/java/com/example/dessertclicker/ui/DessertClickerState.kt
@@ -1,0 +1,10 @@
+/**
+ * uiの状態を管理するクラス
+ */
+data class DessertClickerState(
+    val revenue: Int = 0,
+    val dessertsSold: Int = 0,
+    val currentDessertIndex: Int = 0,
+    val currentDessertPrice: Int = 0,
+    val currentDessertImageId: Int = 0
+)
